% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{boot_FitSparsePCA}
\alias{boot_FitSparsePCA}
\title{boot_FitSparsePCA}
\usage{
boot_FitSparsePCA(Y, B, r, minV = 1, cores = 1, unique = TRUE)
}
\arguments{
\item{Y}{data}

\item{B}{number of bootstrap samples}

\item{r}{number of missing actors}

\item{minV}{minimum number of neighbors of missing actors}

\item{cores}{number of cores for possible parallel computation}

\item{unique}{should unique results only be displayed ?}
}
\description{
Finds initial cliques using a spca on bosstraps sub-samples
}
