% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{comp_Omega}
\alias{comp_Omega}
\title{comp_Omega}
\usage{
comp_Omega(Pg, Rho, p)
}
\arguments{
\item{Pg}{Edge probability matrix}

\item{Rho}{Correlation matrix}

\item{p}{number of observed species}
}
\value{
the market matrix filled with all the precision values common to all spanning trees
}
\description{
Update of the precision terms
}
