% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{complement_spca}
\alias{complement_spca}
\title{Select the k first components and their complement as initial cliques}
\usage{
complement_spca(Y, k)
}
\arguments{
\item{Y}{count dataset}

\item{k}{number of principal components of sparse PCA to keep}
}
\value{
a list of 2*k cliques
}
\description{
This function aims at efficiently exploring the space of likely cliques when only one missing actor is estimated.
}
\examples{
data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=TRUE)
complement_spca(data$Y,k=2)
}
