% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{computeOmega}
\alias{computeOmega}
\title{Updates the precision terms}
\usage{
computeOmega(Pg, Rho, p)
}
\arguments{
\item{Pg}{Edge probability matrix}

\item{Rho}{Correlation matrix}

\item{p}{number of observed species}
}
\value{
the market matrix filled with all the precision values common to all spanning trees
}
\description{
Updates the precision terms
}
