% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exact_comp.R
\name{det.fractional}
\alias{det.fractional}
\title{Exact log-determinant computation}
\usage{
det.fractional(A, log = TRUE)
}
\arguments{
\item{A}{a square positive definite matrix}

\item{log}{should result be in log scale}
}
\value{
the exact determinant of A
}
\description{
Exact log-determinant computation
}
