% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{exactMeila}
\alias{exactMeila}
\title{Calculates the Meila matrix using exact computation}
\usage{
exactMeila(W, r)
}
\arguments{
\item{W}{A weight matrix.}

\item{r}{The number of missing actors.}
}
\value{
The Meila matrix.
}
\description{
Calculates the Meila matrix using exact computation
}
