% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions-V2.R
\name{FitSparsePCA}
\alias{FitSparsePCA}
\title{FitSparsePCA
Fit sparse PCA on a grid of alpha}
\usage{
FitSparsePCA(Y, r = 1, minV = 1, alphaGrid = 10^(seq(-4, 0, by =
  0.1)))
}
\arguments{
\item{Y}{a data frame}

\item{r}{the number of missing actors}

\item{minV}{the minimum number of neighbors for each missing actor}

\item{alphaGrid}{the grid for alpha}
}
\value{
a list
}
\description{
FitSparsePCA
Fit sparse PCA on a grid of alpha
}
\examples{
#blabla
}
