% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{init_blockmodels}
\alias{init_blockmodels}
\title{init_blockmodels}
\usage{
init_blockmodels(k, counts, sigma_obs, MO, SO, alpha = 0.1)
}
\arguments{
\item{k}{number of groups}

\item{counts}{data}

\item{sigma_obs}{original covariance matrix estimate}

\item{MO}{original observed means estimate}

\item{SO}{original observed marginal variances estimate}

\item{alpha}{tempering parameter}
}
\value{
a list of possible cliques
}
\description{
Find initial cliques using blockmodels on the initial network (possibly inferred using EMtree or VEMtree with r=0)
}
