% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{initOmega}
\alias{initOmega}
\title{Initialize Sigma and Omega using an initial clique of neighbors of the missing actor}
\usage{
initOmega(Sigma = NULL, cliqueList, cst = 1.1)
}
\arguments{
\item{Sigma}{variance-covariance matrix (p x p)}

\item{cliqueList}{list of found cliques, given as vectorss of nodes indices}

\item{cst}{small constant for positive definiteness}
}
\value{
\itemize{
\item{Sigma0: }{initial value of complete covariance matrix ((p+r)x(p+r) matrix)}
\item{K0: }{initial value of complete precision matrix ((p+r)x(p+r) matrix)}
\item{clique: }{vector containing the indices of the nodes in the clique}}
}
\description{
Initialize Sigma and Omega by taking the principal component of cliques as initial value for the hidden variables.
The PCA is done on the correaltion matrix, the variance of hidden variables are set to the empirical variances of the pca.
The corresponding precision term in omega is set to 1 for identifiability reasons.
}
\examples{
data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=TRUE)
Sigma=data$Sigma
initClique=FitSparsePCA(data$Y,r=1)$cliques
initOmega(Sigma=Sigma, cliqueList=initClique)
}
