% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{initVEM}
\alias{initVEM}
\title{Initialize all parameters for the variational inference}
\usage{
initVEM(cliqueList, sigma_O, MO, r)
}
\arguments{
\item{cliqueList}{List of size r of initial neighbors for each missing actor.}

\item{sigma_O}{PLNmodels output: estimated observed bloc of the variance-covariance matrix.}

\item{MO}{PLNmodels output: estimated variational mean values of the latent parameters corresponding to observed species.}

\item{r}{Number of missing actors.}
}
\value{
Computes the following initial matrices:
\itemize{
\item{Wginit:}{ variational edges weights matrix.}
\item{Winit:}{ edges weights matrix.}
\item{omegainit:}{ precision matrix.}
\item{MHinit:}{ variational mean values for the hidden latent Gaussian parameters.}}
}
\description{
Initialize all parameters for the variational inference
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=TRUE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
sigma_O=PLNfit$sigma_O
#-- find initial clique using your favorite method
findclique=FitSparsePCA(MO,r=1)
initClique=findclique$cliques
#-- initialize the VEM
initList=initVEM(cliqueList=initClique,sigma_O=sigma_O, MO=MO,r=1 )
str(initList)
}
