% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{Kirshner}
\alias{Kirshner}
\title{Computes edges probability from weights W (Kirshner (07) formulas)}
\usage{
Kirshner(W, r, it1, verbatim = 0)
}
\arguments{
\item{W}{A weight matrix.}

\item{r}{Number of missing actors.}

\item{it1}{Checks if nestorFit is at it first iteration.}

\item{verbatim}{Displays unstable probabilities if set to 2.}
}
\value{
The matrix of edges probabilities.
}
\description{
Computes edges probability from weights W (Kirshner (07) formulas)
}
