% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{List.nestorFit}
\alias{List.nestorFit}
\title{Run function nestorFit on a list of initial cliques using parallel computation (mclapply)}
\usage{
List.nestorFit(
  cliqueList,
  Y,
  sigma_O,
  MO,
  SO,
  r,
  alpha = 0.1,
  cores = 1,
  maxIter = 20,
  eps = 0.001,
  trackJ = FALSE
)
}
\arguments{
\item{cliqueList}{a list containing all initial cliques to be tested}

\item{Y}{count dataset}

\item{sigma_O}{result of PLN estimation: variance covariance matrix of observed data}

\item{MO}{result of PLN estimation: means matrix of observed data}

\item{SO}{result of PLN estimation: marginal variances matrix of observed data}

\item{r}{number of hidden variables}

\item{alpha}{tempering parameter}

\item{cores}{number of cores}

\item{maxIter}{maximal number of iterations of the algorithm}

\item{eps}{convergence precision parameter}

\item{trackJ}{boolean for the estimation of the lower bound at each parameter update}
}
\value{
a list containing the fit of nestorFit for every cliques contained in cliqueList
}
\description{
Run function nestorFit on a list of initial cliques using parallel computation (mclapply)
}
\examples{
 data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- find a list of initial cliques
findcliqueList=boot_FitSparsePCA(data$Y, B=5, r=1)
cliqueList=findcliqueList$cliqueList
length(cliqueList)
#-- run List.nestorFit
fitList=List.nestorFit(cliqueList,data$Y, sigma_O, MO,SO, r=1)
length(fitList)
str(fitList[[1]])
}
