% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{logSumTree}
\alias{logSumTree}
\title{Exact computation of matrix tree log determinant}
\usage{
logSumTree(W)
}
\arguments{
\item{W}{A weight matrix.}
}
\value{
\itemize{
\item{det:}{ exact log-determinant of W Laplacian matrix.}
\item{max.prec:}{ boolean tracking the reach of maximal precision during computation.}}
}
\description{
Exact computation of matrix tree log determinant
}
