% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{missing_from_scratch}
\alias{missing_from_scratch}
\title{missing_from_scratch}
\usage{
missing_from_scratch(n, p, r, type, plot = FALSE, dens = 2/p)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of observed species}

\item{r}{number of missing actors}

\item{type}{type of conditional dependency structure (graph), either "erdos", "cluster", or "scale-free".}

\item{plot}{should the simulated graph be plotted ? default to FALSE}

\item{dens}{density parameter for cluster and erdos graphs. For erdos graphs, this corresponds to the edges probability of connectance}
}
\value{
\itemize{
\item{Y:}{}
\item{UH:}{}
\item{Sigma:}{}
\item{Omega:}{}
\item{G:}{}
\item{TC:}{}
\item{H:}{}}
}
\description{
Function to simulate data and parameters
}
