% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{Mstep}
\alias{Mstep}
\title{Computes the maximization step of the algorithm}
\usage{
Mstep(M, S, Pg, Omega, W, Wg, p, logSTW, logSTWg, trackJ = FALSE)
}
\arguments{
\item{M}{Matrix of variational means.}

\item{S}{Matrix of variational marginal variances.}

\item{Pg}{Matrix of edges probabilities.}

\item{Omega}{Matrix gathering precision terms common to all spanning tree structures.}

\item{W}{Edges weights matrix.}

\item{Wg}{Variational edge weights matrix.}

\item{p}{Number of observed species.}

\item{logSTW}{Log of the Matrix Tree quantity of the W matrix.}

\item{logSTWg}{Log of the Matrix Tree quantity of the Wg matrix.}

\item{trackJ}{Boolean for evaluating the lower bound at each parameter update.}
}
\value{
Quantities required by the VEstep function:
\itemize{
\item{W:}{ edge weights matrix.}
\item{Omega:}{ Matrix gathering precision terms common to all spanning tree structures.}
\item{LB:}{ lower bound values.}
\item{logSTW:}{ log of the Matrix Tree quantity of the W matrix.}
\item{max.prec:}{ boolean tracking the reach of maximal precision during computation.}}
}
\description{
Computes the maximization step of the algorithm
}
