% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{nestor}
\alias{nestor}
\title{Core function of nestor.}
\usage{
nestor(Y, MO, SO, initList, maxIter = 20, eps = 0.01, alpha = 0.1,
  verbatim = 1, print.hist = FALSE, trackJ = FALSE)
}
\arguments{
\item{Y}{count dataset}

\item{MO}{estimated means from norm_PLN}

\item{SO}{estimated marginal vairances from norm_PLN}

\item{initList}{result from initVEM}

\item{maxIter}{maximun number of iterations}

\item{eps}{convergence precision parameter}

\item{alpha}{tempering parameter, default to 0.1}

\item{verbatim}{controls verbosity, between 0,1 and 2.}

\item{print.hist}{prints edges weights histograms at each step if TRUE}

\item{trackJ}{computes the lower bound at each parameter update if TRUE. Otherwise, the lower bound is only computed at each new VE step.}
}
\value{
\itemize{
\item{M}{ estimated means}
\item{S}{ estimated marginal variances}
\item{Pg}{ edges probabilities}
\item{Wg}{edges variational weights}
\item{W}{ edges weights}
\item{Omega}{ market matrix filled with precision terms common to all spanning trees}
\item{lowbound}{ table containing the lowerbound trajectory}
\item{features}{ table containing the parametes trajectory}
\item{finalIter}{ number of iterations until convergence was reached}
\item{time}{ running time of the VEM}
\item{max.prec}{ boolean for reach of maximal precision reached during the VEM fit}}
}
\description{
Core function of nestor.
}
\examples{
data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_obs=PLNfit$sigma_obs
#-- initialize with true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_obs, MO,r=1 )
#-- run core function nestor
with_trackJ=nestor(data$Y, MO,SO, initList=initList, maxIter=5)
}
