% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{nestorFit}
\alias{nestorFit}
\title{Core function of nestorFit}
\usage{
nestorFit(
  Y,
  MO,
  SO,
  initList,
  maxIter = 20,
  eps = 0.01,
  alpha = 0.1,
  verbatim = 1,
  print.hist = FALSE,
  trackJ = FALSE
)
}
\arguments{
\item{Y}{Count dataset.}

\item{MO}{Estimated means from norm_PLN.}

\item{SO}{Estimated marginal vairances from norm_PLN.}

\item{initList}{Result list from initVEM.}

\item{maxIter}{Maximal number of iterations.}

\item{eps}{Convergence precision parameter.}

\item{alpha}{Tempering parameter, default to 0.1.}

\item{verbatim}{Integer controlling verbosity in three levels, starting at 0.}

\item{print.hist}{Prints edges weights histograms at each step if TRUE.}

\item{trackJ}{Computes the lower bound at each parameter update if TRUE. Otherwise, the lower bound is only computed at each new VE step.}
}
\value{
\itemize{
\item{M:}{ estimated means.}
\item{S:}{ estimated marginal variances.}
\item{Pg:}{ edges probabilities.}
\item{Wg:}{ edges variational weights.}
\item{W:}{ edges weights.}
\item{Omega:}{ matrix filled with precision terms common to all spanning trees.}
\item{lowbound:}{ table containing the lowerbound trajectory.}
\item{features:}{ table containing the parametes trajectory.}
\item{finalIter:}{ number of iterations until convergence was reached.}
\item{time:}{ running time of the VEM.}
\item{max.prec:}{ boolean for reach of maximal precision reached during the VEM fit.}}
}
\description{
Core function of nestorFit
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- initialize with true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_O, MO,r=1 )
#-- run core function nestorFit
fit=nestorFit(data$Y, MO,SO, initList=initList, maxIter=5,verbatim=1)
str(fit)
}
