% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{norm_PLN}
\alias{norm_PLN}
\title{Runs PLN function from the PLNmodels package and normalized the outputs}
\usage{
norm_PLN(Y, X = NULL, O = NULL)
}
\arguments{
\item{Y}{Count dataset (n x p).}

\item{X}{Matrix of covariates (n x d).}

\item{O}{Matrix of offsets (n x p).}
}
\value{
\itemize{
\item{MO:}{ normalized means (n x p).}
\item{SO:}{ normalized marginal variances (n x p).}
\item{sigma_O:}{ correlation matrix (p x p) computed from the variance-covariance matrix estimated by PLNmodels, and corresponding to observed variables.}
\item{theta:}{ matrix of the covariates regression coefficients (p x (d+1) including the intercept).}}
}
\description{
Runs PLN function from the PLNmodels package and normalized the outputs
}
\examples{
 n=100 # 100 samples
p=10  # 10 species
r=1 # 1 missing actor
data=generate_missing_data(n=n,p=p,r=r,type="scale-free", plot=TRUE)
X=data.frame(X1=rnorm(n), X2=runif(n))
normPLNfit<-norm_PLN(data$Y,X)
str(normPLNfit)
}
