% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exact_comp.R
\name{pivot.fractional}
\alias{pivot.fractional}
\title{Exact computation of the determinant using the Gauss pivot method}
\usage{
pivot.fractional(A, trace = FALSE, inverse = TRUE)
}
\arguments{
\item{A}{a square positive definite matrix}

\item{trace}{verbosity}

\item{inverse}{compute inverse matrix}
}
\value{
\itemize{
\item{A}{ the original matrix}
\item{det.factors}{ factors of the determinant}
\item{A.inv}{ inverse of A if computed}}
}
\description{
Exact computation of the determinant using the Gauss pivot method
}
