% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{plotConv}
\alias{plotConv}
\title{Plot function for nestor convergence}
\usage{
plotConv(nestorFit)
}
\arguments{
\item{nestorFit}{a fit from the nestor() function}
}
\value{
visualiaztion of nestor convergence
}
\description{
Plot function for nestor convergence
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- use true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_O, MO,r=1 )
nestorFit=nestor(data$Y, MO,SO, initList=initList, maxIter=3 , verbatim=1)
#-- obtain criteria
plotConv(nestorFit)
}
