% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEMtree-package.R
\docType{package}
\name{VEMtree-package}
\alias{VEMtree}
\alias{VEMtree-package}
\title{VEMtree: Network inference with missing actors}
\description{
VEMtree runs a variational inference using an average on spanning trees to infer the conditional dependency structure from abundance data, while accounting for missing actors.
}
\author{
\strong{Maintainer}: Raphaelle Momal \email{rmomaleisenring@gmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
