% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{alphaMax}
\alias{alphaMax}
\title{Heuristic for an upper value of tempering parameter alpha}
\usage{
alphaMax(q, n, sup_val = 0.8, delta = .Machine$double.xmax)
}
\arguments{
\item{q}{Total number of variables (observed and unobserved).}

\item{n}{Number of samples.}

\item{sup_val}{maximal value of the complete estimated correlation matrix (default 0.8).}

\item{delta}{Maximal value allowed for a determinant (default maximal machine precision).}
}
\value{
An upper value of tempering parameter alpha.
}
\description{
Heuristic for an upper value of tempering parameter alpha
}
\examples{
q=15
n=50
alphaMax(q,n)
}
