% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{boot_FitSparsePCA}
\alias{boot_FitSparsePCA}
\title{Finds initial cliques using a sparse PCA on bootstraps sub-samples}
\usage{
boot_FitSparsePCA(Y, B, r, minV = 1, cores = 1, unique = TRUE)
}
\arguments{
\item{Y}{data}

\item{B}{number of bootstrap samples}

\item{r}{number of missing actors}

\item{minV}{minimum number of neighbors of missing actors}

\item{cores}{number of cores for possible parallel computation}

\item{unique}{should unique results only be displayed ?}
}
\value{
\itemize{
\item{cliqueList:}{ a list of all possible initial cliques of neighbors. Each element is of size r}
\item{nb_occ:}{ vector of the number of times each cliques has been found by sPCA.}}
}
\description{
Finds initial cliques using a sparse PCA on bootstraps sub-samples
}
\examples{
data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=TRUE)
boot_FitSparsePCA(data$Y, B=20, r=1)
}
