% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{computeWg}
\alias{computeWg}
\title{computeWg}
\usage{
computeWg(Rho, Omega, W, r, n, alpha, hist = FALSE, verbatim = FALSE)
}
\arguments{
\item{Rho}{Correlation matrix}

\item{Omega}{Market matrix filled with all the precision values common to all spanning trees}

\item{W}{Edge weight matrix}

\item{r}{number of missing actors}

\item{n}{number of samples}

\item{alpha}{tempering parameter}

\item{hist}{should the histogram of the log-values of OO blocs and OH blocs be printed ?}

\item{verbatim}{controls verbosity}
}
\value{
The variational edge weights matrix
}
\description{
Central function for the update of the variational edges weights.
}
