% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{exactMeila}
\alias{exactMeila}
\title{Calculates the Meila matrix using exact computation}
\usage{
exactMeila(W, r)
}
\arguments{
\item{W}{a weight matrix}

\item{r}{the number of missing actors}
}
\value{
the Meila matrix
}
\description{
Calculates the Meila matrix using exact computation
}
