% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{ggimage}
\alias{ggimage}
\title{ggplot heatmap}
\usage{
ggimage(data)
}
\arguments{
\item{data}{input matrix}
}
\value{
a heatmap
}
\description{
ggplot heatmap
}
\examples{
Sigma=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)$Sigma
ggimage(Sigma)
}
