% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{ggimage}
\alias{ggimage}
\title{ggplot heatmap}
\usage{
ggimage(data, no.names = FALSE, order = NULL)
}
\arguments{
\item{data}{Input matrix.}

\item{no.names}{Boolean controlling the display of variable names and ticks.}

\item{order}{Defines a specific display order of the heatmap's rows.}
}
\value{
A heatmap of the input data
}
\description{
ggplot heatmap
}
\examples{
Sigma=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)$Sigma
ggimage(Sigma)
}
