% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{init_blockmodels}
\alias{init_blockmodels}
\title{Find initial cliques using blockmodels on the initial marginalized network}
\usage{
init_blockmodels(sigma_O, MO, SO, k = 3, alpha = 0.1, cores = 1)
}
\arguments{
\item{sigma_O}{PLNmodels output: covariance matrix estimate.}

\item{MO}{PLNmodels output: observed means estimate.}

\item{SO}{PLNmodels output: observed marginal variances estimate.}

\item{k}{Number of groups to find.}

\item{alpha}{Tempering parameter.}

\item{cores}{Number of cores for parallel computation (uses mclapply, not available for Windows).}
}
\value{
A list of cliques
}
\description{
This function aims at clustering the marginal network, inferred from edges probabilities obtained with either \code{nestorFit()} or \code{EMtree()}.
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=TRUE)
data$TC
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- initialize with blockmodels
init_blockmodels(sigma_O, MO, SO, k=2 )
}
