% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{init_blockmodels}
\alias{init_blockmodels}
\title{Find initial cliques using blockmodels on the initial marginalized network}
\usage{
init_blockmodels(Y, sigma_O, MO, SO, k = 3, poisson = FALSE,
  alpha = 0.1, cores = 1)
}
\arguments{
\item{Y}{count data}

\item{sigma_O}{original covariance matrix estimate}

\item{MO}{original observed means estimate}

\item{SO}{original observed marginal variances estimate}

\item{k}{number of groups}

\item{poisson}{boolean for the choice of model of blockmodel. If FALSE, runs bernoulli.}

\item{alpha}{tempering parameter}

\item{cores}{number of cores}
}
\value{
a list of possible cliques
}
\description{
Find initial cliques using blockmodels on the initial marginalized network
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=TRUE)
data$TC
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- initialize with blockmodels
init_blockmodels(data$Y,sigma_O, MO, SO, k=2 )
}
