% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{init_mclust}
\alias{init_mclust}
\title{Find initial cliques using mclust on the estimated correlation matrix}
\usage{
init_mclust(Sigma, r, n.noise = NULL)
}
\arguments{
\item{Sigma}{Estimated correlation matrix from observed counts (pxp)}

\item{r}{number of missing actors}

\item{n.noise}{quantity of noise for mclust, set to 3*p by default.}
}
\value{
A list of size r, with initial cliques of neighbors for each missing actor.
}
\description{
Find initial cliques using mclust on the estimated correlation matrix
}
\examples{
data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=TRUE)
PLNfit<-norm_PLN(data$Y)
Sigma_hat=PLNfit$sigma_obs
#-- original true clique
data$TC
#-- clique found by mclust
clique_mclust=init_mclust(Sigma_hat, r=1)
clique_mclust
}
