% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{initVEM}
\alias{initVEM}
\title{initVEM}
\usage{
initVEM(counts, cliquelist, sigma_obs, MO, r)
}
\arguments{
\item{counts}{count data matrix}

\item{cliquelist}{list of initial neighbors for the missing actors}

\item{sigma_obs}{estimated observed bloc of the variance-covariance matrix}

\item{MO}{estimated mean values of the latent parameters corresponding to observed species}

\item{r}{number of missing actors}
}
\value{
initVEM computes the following initial matrices:
\itemize{
\item{Wginit:}{ Variational edges weights matrix}
\item{Winit:}{ Edges weights matrix}
\item{omegainit:}{ Precision matrix}
\item{MHinit:}{ Mean values for the hidden latent Gaussian parameters}}
}
\description{
wraper to initialize all parameters for the variational inference
}
