% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{initVEM}
\alias{initVEM}
\title{initVEM}
\usage{
initVEM(counts, cliquelist, sigma_obs, MO, r)
}
\arguments{
\item{r}{}
}
\description{
wraper to initialize all parameters for the variational inference
}
