% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{initVEM}
\alias{initVEM}
\title{Initialize all parameters for the variational inference}
\usage{
initVEM(Y, cliqueList, sigma_obs, MO, r)
}
\arguments{
\item{Y}{count data matrix}

\item{cliqueList}{list of size r of initial neighbors for each missing actor}

\item{sigma_obs}{estimated observed bloc of the variance-covariance matrix}

\item{MO}{estimated mean values of the latent parameters corresponding to observed species}

\item{r}{number of missing actors}
}
\value{
initVEM computes the following initial matrices:
\itemize{
\item{Wginit:}{ Variational edges weights matrix}
\item{Winit:}{ Edges weights matrix}
\item{omegainit:}{ Precision matrix}
\item{MHinit:}{ Mean values for the hidden latent Gaussian parameters}}
}
\description{
Initialize all parameters for the variational inference
}
\examples{
data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=TRUE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
sigma_obs=PLNfit$sigma_obs
#-- find initial clique
findclique=FitSparsePCA(data$Y,r=1)
initClique=findclique$cliques
#-- initialize the VEM
initVEM(Y=data$Y,cliqueList=initClique,sigma_obs=sigma_obs, MO=MO,r=1 )
}
