% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{Kirshner}
\alias{Kirshner}
\title{Computes edges probability from weights W (Kirshner (07) formulas)}
\usage{
Kirshner(W, r, it1, verbatim = 0)
}
\arguments{
\item{W}{a weight matrix}

\item{r}{number of missing actors}

\item{it1}{check if nestorFit is at it first iteration}

\item{verbatim}{boolean controling verbosity}
}
\value{
matrix of edges probabilities
}
\description{
Computes edges probability from weights W (Kirshner (07) formulas)
}
