% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{List.nestorFit}
\alias{List.nestorFit}
\title{Run function nestorFit on a list of initial cliques using parallel computation (mclapply)}
\usage{
List.nestorFit(
  cliqueList,
  Y,
  sigma_O,
  MO,
  SO,
  r,
  alpha = 0.1,
  cores = 1,
  maxIter = 20,
  eps = 0.001,
  trackJ = FALSE
)
}
\arguments{
\item{cliqueList}{List containing all initial cliques to be tested.}

\item{Y}{Count dataset.}

\item{sigma_O}{Result of PLN estimation: variance covariance matrix of observed data.}

\item{MO}{Result of PLN estimation: means matrix of observed data.}

\item{SO}{Result of PLN estimation: marginal variances matrix of observed data.}

\item{r}{Number of hidden variables.}

\item{alpha}{Tempering parameter.}

\item{cores}{Number of cores for parallel computation (uses mclapply, not available for Windows).}

\item{maxIter}{Maximal number of iterations of the algorithm.}

\item{eps}{Convergence precision parameter.}

\item{trackJ}{Boolean for the lower bound estimation at each parameter update instead of each step.}
}
\value{
A list containing the fit of nestorFit for every clique contained in cliqueList.
}
\description{
Run function nestorFit on a list of initial cliques using parallel computation (mclapply)
}
\examples{
 data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- find a list of initial cliques
findcliqueList=boot_FitSparsePCA(MO, B=5, r=1)
cliqueList=findcliqueList$cliqueList
length(cliqueList)
#-- run List.nestorFit
fitList=List.nestorFit(cliqueList,data$Y, sigma_O, MO,SO, r=1)
length(fitList)
str(fitList[[1]])
}
