% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{logSumTree}
\alias{logSumTree}
\title{exact computation of matrix tree log determinant}
\usage{
logSumTree(W)
}
\arguments{
\item{W}{a weight matrix}
}
\value{
\itemize{
\item{det}{ exact log-determinant}
\item{max.prec}{ boolean tracking the reach of maximal precision during computation}}
}
\description{
exact computation of matrix tree log determinant
}
