% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{LowerBound}
\alias{LowerBound}
\title{Computes the lower bound.}
\usage{
LowerBound(Pg, Omega, M, S, W, Wg, p, logSTW, logSTWg)
}
\arguments{
\item{Pg}{Edge probability matrix (qxq)}

\item{Omega}{market matrix with precision values (qxq)}

\item{M}{Matrix of variational means (nxq)}

\item{S}{Matrix of variational marginal variances (nxq)}

\item{W}{Edges weight matrix}

\item{Wg}{Variational edges weight matrix}

\item{p}{number of observed species}

\item{logSTW}{log of the matrix tree run on the W matrix}

\item{logSTWg}{log of the matrix tree run on the Wg matrix}
}
\value{
a vector containing the different terms of the lower bound
}
\description{
Computes the lower bound.
}
