% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{missing_from_scratch}
\alias{missing_from_scratch}
\title{missing_from_scratch}
\usage{
missing_from_scratch(n, p, r = 1, type, plot = FALSE, dens = 2/p)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of observed species}

\item{r}{number of missing actors}

\item{type}{type of conditional dependency structure (graph), either "erdos", "cluster", or "scale-free".}

\item{plot}{should the simulated graph be plotted ? default to FALSE}

\item{dens}{density parameter for cluster and erdos graphs. For erdos graphs, this corresponds to the edges probability of connectance}
}
\value{
\itemize{
\item{Y:}{ Observed count data}
\item{UH:}{ Hidden latent Gaussian parameters}
\item{Sigma:}{ Observed bloc of the variance-covariance matrix}
\item{Omega:}{ Observed bloc of the precision matrix}
\item{G:}{ The generated conditional dependency structure}
\item{TC:}{ The true clique of neighbors of the missing actors (a list if r>1)}
\item{H:}{  Indexes corresponding to missing actors in the original data}}
}
\description{
Function to simulate data and parameters including missing actors. Setting r=1 is advised.
}
