% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{Mstep}
\alias{Mstep}
\title{Title}
\usage{
Mstep(M, S, Pg, Omega, W, Wg, p, logSTW, logSTWg, trackJ = FALSE)
}
\arguments{
\item{M}{matrix of means}

\item{S}{matrix of marginal variances}

\item{Pg}{matrix of edges probabilities}

\item{Omega}{matrix gathering the precision terms}

\item{W}{edges weights matrix}

\item{Wg}{variational weights matrix}

\item{p}{number of observed species}

\item{logSTW}{log-sum tree of the W matrix}

\item{logSTWg}{log-sum tree of the Wg matrix}

\item{trackJ}{boolean for evaluating the lower bound at each parameter update}
}
\description{
Title
}
