% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{norm_PLN}
\alias{norm_PLN}
\title{norm_PLN}
\usage{
norm_PLN(Y)
}
\arguments{
\item{Y}{count dataset}
}
\value{
\itemize{
\item{MO}{ Normalized means}
\item{SO}{ Normalized marginal variances}
\item{sigma_obs}{Vairance-covariance matrix Sigma as estimated by PLNmodels}}
}
\description{
Runs PLN function from the PLNmodels package and normalized the outputs
}
\examples{
 data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=TRUE)
norm_PLN(data$Y)
}
