% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{plotConv}
\alias{plotConv}
\title{Plot function for nestorFit convergence}
\usage{
plotConv(nestorFit)
}
\arguments{
\item{nestorFit}{Fit from the nestorFit() function.}
}
\value{
Visualiaztion of nestorFit convergence.
}
\description{
Plot function for nestorFit convergence
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- use true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_O, MO,r=1 )
nestorFit=nestorFit(data$Y, MO,SO, initList=initList, maxIter=3 , verbatim=1)
#-- obtain criteria
plotConv(nestorFit)
}
