% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{plotPerf}
\alias{plotPerf}
\title{Comparaitve plot of estimated edge probabilities and original graph.}
\usage{
plotPerf(P, G, r, thresh = 0.5, no.names = FALSE)
}
\arguments{
\item{P}{Edges probabilities matrix.}

\item{G}{Adjacency matrix of the original graph.}

\item{r}{Number of missing actors.}

\item{thresh}{Required threshold for criteria computations, default to 0.5.}

\item{no.names}{Boolean controlling the display of variable names and ticks.}
}
\value{
Two heatmaps with performance criteria computed for the specified threshold as title.
}
\description{
Comparaitve plot of estimated edge probabilities and original graph.
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- use true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(cliqueList=initClique,sigma_O, MO,r=1 )
nestorFit=nestorFit(MO,SO, initList=initList, maxIter=3,verbatim=1 )
#-- obtain criteria
plotPerf(nestorFit$Pg, data$G,r=1)
}
