% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{plotPerf}
\alias{plotPerf}
\title{Comparaitve plot of estimated edge probabilities and original graph.}
\usage{
plotPerf(P, G, r, thresh = 0.5)
}
\arguments{
\item{P}{edges probabilities matrix}

\item{G}{adjacency matrix of the original graph}

\item{r}{number of missing actors}

\item{thresh}{required threshold for criteria computations, default to 0.5}
}
\value{
two heatmaps with performance criteria computed for the specified threshold as title.
}
\description{
Comparaitve plot of estimated edge probabilities and original graph.
}
\examples{
data=missing_from_scratch(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_obs=PLNfit$sigma_obs
#-- use true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_obs, MO,r=1 )
nestorFit=nestor(data$Y, MO,SO, initList=initList, maxIter=3 )
#-- obtain criteria
plotPerf(nestorFit$Pg, data$G,r=1)
}
