% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{ppvtpr}
\alias{ppvtpr}
\title{Computes precision and recall statistics separating observed from hidden variables, and (FPR,FNR) for hidden variables .}
\usage{
ppvtpr(probs, G, r, thresh = 0.5)
}
\arguments{
\item{probs}{matrix of estimated edges probabilities}

\item{G}{original graph}

\item{r}{number of missing actors}

\item{thresh}{required threshold for criteria computations, default to 0.5}
}
\value{
\itemize{
\item{PPV}{ Precision of the whole data}
\item{PPVH}{ Precision regarding observed data}
\item{PPVO}{ Precision regarding hidden data}
\item{TPR}{ Recall of the whole data}
\item{TPRH}{ Recall regarding observed data}
\item{TPRO}{ Recall regarding hidden data}
\item{FPRH}{ False Positive Rate of hidden data}
\item{FNRH}{ False Negative Rate of hidden data}
}
}
\description{
Computes precision and recall statistics separating observed from hidden variables, and (FPR,FNR) for hidden variables .
}
\examples{
data=generate_missing_data(n=100,p=10,r=1,type="scale-free", plot=FALSE)
PLNfit<-norm_PLN(data$Y)
MO<-PLNfit$MO
SO<-PLNfit$SO
sigma_O=PLNfit$sigma_O
#-- use true clique for example
initClique=data$TC
#-- initialize the VEM
initList=initVEM(data$Y,cliqueList=initClique,sigma_O, MO,r=1 )
nestorFit=nestor(data$Y, MO,SO, initList=initList, maxIter=3,verbatim=1 )
#-- obtain criteria
ppvtpr(nestorFit$Pg,r=1, data$G)
}
