% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer.R
\name{VEstep}
\alias{VEstep}
\title{Computes the variational expectation step of the algorithm}
\usage{
VEstep(MO, SO, SH, Omega, W, Wg, MH, Pg, logSTW, logSTWg, alpha, it1,
  verbatim, trackJ = FALSE, hist = FALSE)
}
\arguments{
\item{MO}{matrix of observed means}

\item{SO}{matrix of observed marginal variances}

\item{SH}{matrix of observed hidden variances}

\item{Omega}{matrix containing the precision terms of precision matrices faithful ot a tree}

\item{W}{edges weights matrix}

\item{Wg}{variational edges weights matrix}

\item{MH}{matrix of hidden means}

\item{Pg}{edges probabilities matrix}

\item{logSTW}{log-sum tree of the W matrix}

\item{logSTWg}{log-sum tree of the Wg matrix}

\item{alpha}{tempering parameter}

\item{it1}{check if nestor is at its first iteration}

\item{verbatim}{boolean controling verbosity}

\item{trackJ}{boolean for evaluating the lower bound at each parameter update}

\item{hist}{boolean for printing edges weights histogram at each iteration}
}
\description{
Computes the variational expectation step of the algorithm
}
